accounts/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto==2.49.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.10
django-admin-interface==0.18.3
django-allauth==0.47.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-bootstrap-form==3.4
django-braces==1.15.0
django-ckeditor==6.2.0
django-classy-tags==2.0.0
django-cms==3.9.0
django-colorfield==0.6.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-etc==1.3.1
django-filer==2.1.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-formtools==2.3
django-hitcount==1.3.3
django-ipware==4.0.2
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-meta==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-simple-pagination==1.3
django-sms==0.5.0
django-social-share==2.2.1
django-star-ratings==0.9.2
django-storages==1.12.3
django-summernote==0.8.20.0
django-taggit==2.0.0
django-tracking-analyzer==1.1.1
django-treebeard==4.5.1
django-user-agents==0.4.0
djangocms-admin-style==2.0.2
djangocms-page-meta==1.0.1
easy-thumbnails==2.8
error-tracker==2.1.0
frozenlist==1.2.0
geoip2==4.5.0
idna==3.3
lxml==4.7.1
maxminddb==2.2.0
messagebird==2.0.0
microurl==0.1.1
multidict==5.2.0
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-geoip==1.2
python-geoip-geolite2==2015.303
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.5
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.1.0
swapper==1.3.0
tinycss2==1.1.1
twilio==7.4.0
ua-parser==0.10.0
Unidecode==1.1.2
urllib3==1.26.7
user-agents==2.2.0
webencodings==0.5.1
yarl==1.7.2
